+++ packages/apps/explorer/src/components/CompactTable/utils/formatLink.tsx
@@ -34,12 +34,3 @@ export const FormatLink = ({ appendUrl }: IOptions): FC<IProps> => {
 
   return Component;
 };
-
-export const FormatJsonParse = (): FC<IProps> => {
-  const Component: FC<IProps> = ({ value }) => (
-    <Text variant="code" className={dataFieldClass}>
-      {!!value && value?.length > 0 ? JSON.parse(value) : value}
-    </Text>
-  );
-  return Component;
-};
