--- packages/apps/explorer/src/components/CompactTable/utils/formatStatus.tsx
@@ -1,13 +1,0 @@
-import { MonoCheck, MonoClear } from '@kadena/kode-icons/system';
-import type { FC } from 'react';
-import React from 'react';
-
-interface IProps {
-  value: string;
-}
-
-export const FormatStatus = (): FC<IProps> => {
-  const Component: FC<IProps> = ({ value }) =>
-    value ? <MonoCheck /> : <MonoClear />;
-  return Component;
-};
