+++ packages/libs/kode-ui/src/patterns/CompactTable/TableFormatters/FormatAmount.tsx
