+++ packages/libs/kode-ui/src/patterns/CompactTable/TableFormatters/FormatDefault.tsx
