+++ packages/libs/kode-ui/src/patterns/CompactTable/TableFormatters/FormatLink.tsx
